converted src/views/sign/*
