+++ src/custom.d.ts
@@ -1,0 +1,5 @@
+// for importing file of .svg
+declare module '*.svg' {
+  const content: any;
+  export default content;
+}
