+++ src/views/sign/styles/signStyle.ts
@@ -1,0 +1,28 @@
+export const signStyle ={
+  signin_container: {
+    display: "flex",
+    flexFlow: "column",
+    border: "solid 2px #f0f0f0",
+    borderRadius: "16px",
+  },
+  title: {
+    fontSize: "32px",
+    textAlign: "center",
+    fontWeight: "bold",
+  },
+  sub_title: {
+    fontSize: "24px",
+  },
+  google_button: {
+    fontSize: "14px",
+  },
+  google_logo: {
+    position: "absolute",
+    left: "16px",
+  },
+  border: {
+    borderBottom: " 1px solid #c0c0c0",
+    width: "100%",
+    margin: "0 12px",
+  },
+};
